- [CLN] Renamed geometry to structs - [ENH] Welly is optional - [ENH] structured_elements.py and unstructured_elements.py have been updated - [ENH] subpackage for visualiztion. pyvista code moved there - [DOC] Improved definition of the interfaces subpackage
