+++ subsurface/structs/__init__.py
@@ -1,2 +1,2 @@
 from .unstructured_elements import PointSet, TriSurf, LineSet, TetraMesh, CurvilinearMesh
-from .structured_elements import GridData
+from .structured_elements import StructuredGrid
