+++ subsurface/structs/common.py
@@ -1,7 +1,11 @@
+from typing import Union
+
+from subsurface.geometry.base_structures import UnstructuredData, StructuredData
 
 
 class Common(object):
     """A set of shared functionality for all spatially referenced data."""
+    data: Union[UnstructuredData, StructuredData]
 
     def validate(self):
         raise NotImplementedError()
